Updated README.md to account for ggplot2 updates
ggplot fixed their upstream issue (see tidyverse/ggplot2#4149) but instead of throwing an error there are still some
cases where it uses the default colours as opposed to the set themes ones. So this is just to update the readme to make
this clear.

Until we find what is causing the colour to use defaults I suppose #44 should remain open.